requirements.txt
torch
torchvision
numpy
matplotlib
Pillow
imagecorruptions